Add files via upload
run in python (if you are in a terminal or command prompt I think you can just type in   python -i Gauss.py    ) and play around. I might flesh it in, I might not. For the main reason I made this, you have to have a gaussian integer, say x, and type x.cplot(n) for some natural number n. What this does is take the norm of the G. integer and find all other G. ints such that they share the norm. Then, it takes the norm and either multiplies it by 3 and then adds 1 (this happens if the norm is odd), or divides it by 2 (if it is even). then, we see all the G. ints that have the new norm given by the previous function. This does this n times (Although now that I think about it I might have left out an iteration - oh well). The picture is the collection of all the G. ints throughout the iterations. This uses matplotlib, so I hope it just runs smoothly without any fancy stuff needing to be downloaded to plot anything. I don't know much about python :p